third_party/WebKit/LayoutTests/http/tests/csspaint/parse-secure-context.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/get-host-info.js"></script>
<style>
#test {
  background: red;
  background: paint(success) green;
}
</style>
<div id="test"></div>
<script>
  if (window.location.origin != get_host_info().AUTHENTICATED_ORIGIN) {
    window.location = get_host_info().AUTHENTICATED_ORIGIN + window.location.pathname;
  } else {
    test(function() {
      const GREEN = 'rgb(0, 128, 0)';
      const element = document.getElementById('test');
      assert_true(window.isSecureContext);
      assert_equals(getComputedStyle(element).backgroundColor, GREEN);
    });
  }
</script>
